pyproject.toml
[project]
name = "fotacos"
version = "0.0.1"
description = "Es un app para la rasbery de fotos"
authors = [{ name = "monsignor", email = "monsignor.eduardo@proton.me" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.14,<4.0"

classifiers = [
    "Private :: Do Not Upload",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.1.0",
    "fastapi[standard]>=0.128.0",
    "pillow>=12.1.0",
    "pyside6>=6.10.1",
    "qasync>=0.27.1",
    "tortoise-orm>=0.22.0",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.21",
    "loguru>=0.7.0",
]

[project.scripts]
fotacos = "fotacos.main:main"

[project.urls]
Homepage = "https://MonsignorEduardo.github.io/fotacos/"
Repository = "https://github.com/MonsignorEduardo/fotacos"
Documentation = "https://MonsignorEduardo.github.io/fotacos/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.0",
    "tox-uv>=1.29.0",
    "deptry>=0.24.0",
    "ty>=0.0.1a32",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=1.0.0",
    "honcho>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fotacos"]

[tool.hatch.build.targets.wheel.force-include]
"src/web/dist" = "fotacos/web/dist"

[tool.hatch.build.targets.sdist]
exclude = [
    "src/web/node_modules",
    "src/web/.tanstack",
    "src/web/.vscode",
    "src/web/dist",
]

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
